requirements.txt
forex-python==1.8
google-api-core==2.11.0
google-auth==2.16.1
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-cloud-texttospeech==2.14.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
googletrans==4.0.0rc1
matplotlib==3.7.1
matplotlib-inline==0.1.6
openai==0.27.2
pandas==1.5.3
pycountry==22.3.5
pydub==0.25.1
python-dotenv==1.0.0
pytube==12.1.3
requests==2.28.2
requests-toolbelt==0.10.1
reportlab==3.6.12
streamlit==1.17.0
spacy==3.5.1